+++ include/lomse_screen_drawer.h
@@ -1,6 +1,6 @@
 //---------------------------------------------------------------------------------------
 // This file is part of the Lomse library.
-// Lomse is copyrighted work (c) 2010-2016. All rights reserved.
+// Lomse is copyrighted work (c) 2010-2018. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without modification,
 // are permitted provided that the following conditions are met:
@@ -205,7 +205,7 @@ public:
     void model_point_to_screen(double* x, double* y) const;
 
     //units conversion
-    double Pixels_to_LUnits(Pixels value);
+    LUnits Pixels_to_LUnits(Pixels value);
     Pixels LUnits_to_Pixels(double value);
 
     // settings