+++ include/lomse_xml_parser.h
@@ -84,7 +84,8 @@ public:
     };
 
     //XmlNode helper methods
-    inline bool is_null() { return ! bool(m_node); }
+    inline bool is_null() { return m_node.type() == pugi::xml_node_type::node_null; }
+    //inline bool is_null() { return ! m_node; }      //TODO: is this valid? It seems to work!
 	///get child with the specified name
     inline XmlNode child(const string& name) {
         return XmlNode( m_node.child(name.c_str()));