+++ src/graphic_model/engravers/lomse_rest_engraver.cpp
@@ -1,6 +1,6 @@
 //---------------------------------------------------------------------------------------
 // This file is part of the Lomse library.
-// Lomse is copyrighted work (c) 2010-2016. All rights reserved.
+// Lomse is copyrighted work (c) 2010-2018. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without modification,
 // are permitted provided that the following conditions are met:
@@ -95,8 +95,8 @@ UPoint RestEngraver::get_drag_offset()
     //return center of rest shape
     URect total = m_pRestShape->get_bounds();
     URect rest = m_pRestGlyphShape->get_bounds();
-    return UPoint(rest.get_x() - total.get_x() + rest.get_width() / 2.0,
-                  rest.get_y() - total.get_y() + rest.get_height() / 2.0 );
+    return UPoint(rest.get_x() - total.get_x() + rest.get_width() / 2.0f,
+                  rest.get_y() - total.get_y() + rest.get_height() / 2.0f );
 }
 
 //---------------------------------------------------------------------------------------