+++ src/tests/lomse_test_spacing_algorithm.cpp
@@ -314,42 +314,42 @@ SUITE(SpAlgGourlayTest)
         //check di computation
         list<TimeSlice*>::iterator it = slices.begin();
         MyTimeSlice* pSlice = static_cast<MyTimeSlice*>(*it);
-        CHECK( is_equal_pos(pSlice->get_shortest_duration(), 0.0f) );
+        CHECK( is_equal_time(pSlice->get_shortest_duration(), 0.0) );
         ++it;
         pSlice = static_cast<MyTimeSlice*>(*it);
-        CHECK( is_equal_pos(pSlice->get_shortest_duration(), 16.0f) );
+        CHECK( is_equal_time(pSlice->get_shortest_duration(), 16.0) );
         ++it;
         pSlice = static_cast<MyTimeSlice*>(*it);
-        CHECK( is_equal_pos(pSlice->get_shortest_duration(), 21.333f) );
+        CHECK( is_equal_time(pSlice->get_shortest_duration(), 21.333) );
         ++it;
         pSlice = static_cast<MyTimeSlice*>(*it);
-        CHECK( is_equal_pos(pSlice->get_shortest_duration(), 21.333f) );
+        CHECK( is_equal_time(pSlice->get_shortest_duration(), 21.333) );
         ++it;
         pSlice = static_cast<MyTimeSlice*>(*it);
-        CHECK( is_equal_pos(pSlice->get_shortest_duration(), 21.333f) );
+        CHECK( is_equal_time(pSlice->get_shortest_duration(), 21.333) );
         ++it;
         pSlice = static_cast<MyTimeSlice*>(*it);
-        CHECK( is_equal_pos(pSlice->get_shortest_duration(), 64.0f) );
+        CHECK( is_equal_time(pSlice->get_shortest_duration(), 64.0) );
 
         //check ds computation
         it = slices.begin();
         pSlice = static_cast<MyTimeSlice*>(*it);
-        CHECK( is_equal_pos(pSlice->get_spring_duration(), 0.0f) );
+        CHECK( is_equal_time(pSlice->get_spring_duration(), 0.0) );
         ++it;
         pSlice = static_cast<MyTimeSlice*>(*it);
-        CHECK( is_equal_pos(pSlice->get_spring_duration(), 16.0f) );
+        CHECK( is_equal_time(pSlice->get_spring_duration(), 16.0) );
         ++it;
         pSlice = static_cast<MyTimeSlice*>(*it);
-        CHECK( is_equal_pos(pSlice->get_spring_duration(), 5.333f) );
+        CHECK( is_equal_time(pSlice->get_spring_duration(), 5.333) );
         ++it;
         pSlice = static_cast<MyTimeSlice*>(*it);
-        CHECK( is_equal_pos(pSlice->get_spring_duration(), 21.333f) );
+        CHECK( is_equal_time(pSlice->get_spring_duration(), 21.333) );
         ++it;
         pSlice = static_cast<MyTimeSlice*>(*it);
-        CHECK( is_equal_pos(pSlice->get_spring_duration(), 21.333f) );
+        CHECK( is_equal_time(pSlice->get_spring_duration(), 21.333) );
         ++it;
         pSlice = static_cast<MyTimeSlice*>(*it);
-        CHECK( is_equal_pos(pSlice->get_spring_duration(), 64.0f) );
+        CHECK( is_equal_time(pSlice->get_spring_duration(), 64.0) );
 
 //        cout << test_name() << endl;
 //        ColStaffObjs* pSOCol = pImoScore->get_staffobjs_table();