+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ##### COMPATIBLE CHANGES
 
 - Added new methods for controlling scroll and visual tracking during playback, when using an external player. A new sample 'extplayer' has been added to the 'examples' folder.
+- Fixes for removing Microsoft compiler warnings.